% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathway_namelist.R
\name{get_pathway_namelist}
\alias{get_pathway_namelist}
\title{plot heatmap by integrate data from DESeq and enrichment analysis}
\usage{
get_pathway_namelist(fileName)
}
\arguments{
\item{fileName}{is the post-processed csv file from enrichment analysis}
}
\value{
a list of tibbles of external gene names from different enrichment pathway
}
\description{
more information
}
