% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{perth_table_trans}
\alias{perth_table_trans}
\title{transform a perth mfi excel file}
\usage{
perth_table_trans(data)
}
\arguments{
\item{data}{a excel file}
}
\description{
transform a perth mfi excel file
}
