% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathway_namelist.R
\name{subset_count_maxtrix}
\alias{subset_count_maxtrix}
\title{get tibbles of different subgrouped genes by pathway}
\usage{
subset_count_maxtrix(count_matrix, enrichment)
}
\arguments{
\item{count_matrix}{the RNA-seq expression level count matrix, typically a
normalized count matrix is used}

\item{enrichment_xlsx}{the modified enrichment export xlsx file}

\item{fun}{the fun used to get lists of genes names from enrichment analysis
export}
}
\value{
a list of tibbles of count matrix for different pathways
}
\description{
more information
}
