% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathway_namelist.R
\name{hmplot}
\alias{hmplot}
\title{plot heatmap by integrate data from DESeq and enrichment analysis}
\usage{
hmplot(count_matrix, enrichment, file = "heatmap.pdf", left_all = T)
}
\arguments{
\item{count_matrix}{the RNA-seq expression level count matrix}

\item{file}{the destination and name you want to save your heatmap, by default is the
your current work directory, the default name is heatmap.pdf}

\item{enrichment_xlsx}{the post-processed Excel file from enrichment analysis}
}
\value{
output a pdf file of heatmaps of all the pathways in the enrichment_xlsx
}
\description{
more information
}
